Moved json importing to new file, made import recursive
Import now walks through all JSON files in specified
directory and all subdirectories.
