+++ jsonTransmission/jsonConnect.py
@@ -50,110 +50,6 @@ import datetime
     # (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
     # SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-def doJSONImport():
-    """Read a JSON file and output the file with added values.
-
-    gitHash and version, which are passed as command line arguments,
-    are added to the JSON that is output.
-    """
-    parser = optparse.OptionParser(usage="""\
-                                   %prog [filename] [gitHash] 
-                                   [buildID] [connectionstring]
-                                   [testname] [branch] [platform]""")
-
-    # add in command line options. Add mongo host/port combo later
-    parser.add_option("-f", "--filename", dest="fname",
-                      help="name of file to import",
-                      default=None)
-    parser.add_option("-g", "--githash", dest="ghash",
-                      help="git hash of code being tested",
-                      default=None)
-    parser.add_option("-b", "--buildid", dest="bhash", 
-                      help="build ID of code being tested",
-                      default=None)
-
-    parser.add_option("-c", "--connectionstring", dest="connectstr",
-                      help="string specifying url",
-                      default=None)
-
-    parser.add_option("-t", "--testname", dest="tname",
-                      help="name of the test",
-                      default=None)
-
-    parser.add_option("-r", "--branch", dest="branch",
-                      help="name of the branch",
-                      default=None)
-
-    parser.add_option("-p", "--platform", dest="pform",
-                      help="build platform",
-                      default=None)
-
-
-    (options, args) = parser.parse_args()
-    
-    if options.fname is None:
-        print "\nERROR: Must specify name of file to import\n"
-        sys.exit(-1)
-   
-    if options.ghash is None:
-        print "\nERROR: Must specify git hash \n"
-        sys.exit(-1)
-    
-    if options.bhash is None:
-        print "\nERROR: Must specify build ID \n"
-        sys.exit(-1)
- 
-    if options.connectstr is None:
-        print "\nERROR: Must specify connection string \n"
-        sys.exit(-1)
-   
-    if options.tname is None:
-        print "\nERROR: Must specify test name \n"
-        sys.exit(-1)
-
-    if options.branch is None:
-        print "\nERROR: Must specify branch name \n"
-        sys.exit(-1)
-
-    if options.pform is None:
-        print "\nERROR: Must specify platform \n"
-        sys.exit(-1)
-
-    http_client = tornado.httpclient.HTTPClient()
-
-    for line in open(options.fname, "r"):
-        if line == "\n":
-            continue
-        
-        record = json.loads(line)
-        record["gitHash"] = options.ghash 
-        record["buildID"] = options.bhash 
-        record["testName"] = options.tname
-
-        fileIndex = record["file"].rfind("/") + 1
-        record["dir"] = record["file"][: fileIndex]
-
-        # Add meta info
-        record["meta"] = {}
-        record["meta"]["date"] = str(datetime.datetime.now())
-        record["meta"]["branch"] = options.branch
-        record["meta"]["platform"] = options.pform
-        record["meta"]["gitHash"] = options.ghash 
-        record["meta"]["buildID"] = options.bhash 
-       
-        request = tornado.httpclient.HTTPRequest(
-                                 url=options.connectstr, 
-                                 method="POST", 
-                                 headers={"Content-Type": "application/json"},
-                                 body=json.dumps(record))
-        try:
-            response = http_client.fetch(request)
-            print response.body
-        except tornado.httpclient.HTTPError as e:
-            print "Error: ", e
-    
-    http_client.close()
-
 
 def doJSONCoverage():
     http_client = tornado.httpclient.HTTPClient()
@@ -175,6 +71,7 @@ def doJSONCoverage():
     
     http_client.close()
 
+
 def doJSONAggregate(body):
     http_client = tornado.httpclient.HTTPClient()
     if body == "empty":
@@ -195,6 +92,7 @@ def doJSONAggregate(body):
         print "Error: ", e
     
     http_client.close()
+
 
 def getFileContents():
     configfile = raw_input("Please enter config file name: ")
@@ -223,6 +121,7 @@ def getFileContents():
     http_client.close()
     config.close()
 
+
 class CoverageFormatter(HtmlFormatter):
     def __init__(self):
         HtmlFormatter.__init__(self, linenos="inline")
@@ -243,10 +142,10 @@ class CoverageFormatter(HtmlFormatter):
                 
 
 def main():
-    response = raw_input("Do you want to:\n 1. import 2. aggregate 3. request file " + 
-                         "4. request coverage data\n")
+    response = raw_input("Do you want to:\n 1. request coverage data 2. aggregate " + 
+                         "3. request file\n"
     if response == "1":
-        doJSONImport()
+        doJSONCoverage()
     elif response == "2":
         response = raw_input("Is your GET request:\n 1. empty 2. full \n ")
 
@@ -254,9 +153,7 @@ def main():
             doJSONAggregate("empty")
         else:
             doJSONAggregate("full")
-    elif response == "3":
-        getFileContents()
     else:
-        doJSONCoverage()
+        getFileContents()
 
 main()
